fix: footer module
