+++ src/components/Footer/Footer.module.scss
@@ -1,0 +1,16 @@
+.footer {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 60px;
+  background-color: lightgray;
+  padding: 0 25px;
+
+  &Container {
+    display: flex;
+    max-width: 1200px;
+    justify-content: space-between;
+    width: 100%;
+    margin: 0 auto;
+  }
+}
