+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import styles from './Footer.module.scss';
+
+const Footer = () => {
+  return (
+    <footer className={styles.footer}>
+      <div className={styles.footerContainer}>
+        <div>git1</div>
+        <div>git2</div>
+        <div>git3</div>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
