--- src/components/Footer/footer.scss
@@ -1,8 +1,0 @@
-@import '../../global-styles/variables';
-
-.footer-container {
-  display: flex;
-  justify-content: space-between;
-  background-color: lightgray;
-  padding: 0 40px;
-}
