--- src/components/Footer/footer.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import './footer.scss';
-
-const Footer = () => {
-  return (
-    <div className="footer-container">
-      <div>git1</div>
-      <div>git2</div>
-      <div>git3</div>
-    </div>
-  );
-};
-
-export default Footer;
