+++ src/components/TemplatePage/templatePage.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from '../Header/header';
-import Footer from '../Footer/footer';
+import Header from '../Header/Header';
+import Footer from '../Footer/Footer';
 
 import './templatePage.scss';
